feat: ordered home structure
