+++ frontend/pages/home/components/ICRC/transaction/SendForm/ConfirmDetail.tsx
