+++ frontend/pages/home/components/ICRC/transaction/SendForm/ContactScannerReceiver.tsx
