+++ frontend/pages/home/components/ICRC/transaction/SendForm/ReceiverDetail.tsx
