+++ frontend/pages/home/components/ICRC/transaction/SendForm/ReceiverManual.tsx
