+++ frontend/pages/home/components/ICRC/transaction/SendForm/ReceiverQRScanner.tsx
