+++ frontend/pages/home/components/ICRC/transaction/SendForm/SenderDetail.tsx
