+++ frontend/pages/home/components/ICRC/transaction/SendForm/SenderQRScanner.tsx
