+++ frontend/pages/home/components/ICRC/transaction/TransactionActionBanner.tsx
