+++ frontend/pages/home/components/ICRC/transaction/TransactionInpectDrawer.tsx
